Convert enclosed runner test to new JUnit multiversion test
